requirements_updated.txt
# Core data processing
pandas>=2.2.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Web scraping
selenium>=4.15.0
scrapy>=2.11.0

# Data validation and cleaning
pydantic>=2.5.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging and monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Async and concurrency
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Caching
redis>=5.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools (optional)
jupyter>=1.0.0
ipython>=8.17.0
